refactor(reports): split about-kvk report repository modules (#61)
Move report data access from the monolithic reports repository into file-based About KVK modules with a composer index so report scaling work can add new templates/data sources without growing a single file.
