+++ backend/repositories/reports/index.js
@@ -1,0 +1,5 @@
+const aboutkvkReportRepository = require('./aboutkvkReport/index.js');
+
+module.exports = {
+    ...aboutkvkReportRepository,
+};
